backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
pydantic-settings==2.12.0
asyncpg==0.31.0
psycopg2-binary==2.9.11
pgvector==0.4.2
langchain==1.0.8
langchain-community==0.4.1
google-generativeai==0.8.3 
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
gunicorn==21.2.0
